Move RoadStationRemovalArgs to RemoveRoadStation.h
