+++ src/OpenLoco/src/GameCommands/GameCommands.h
@@ -311,40 +311,6 @@ namespace OpenLoco::GameCommands
         }
     };
 
-    struct RoadStationRemovalArgs
-    {
-        static constexpr auto command = GameCommand::removeRoadStation;
-
-        RoadStationRemovalArgs() = default;
-        explicit RoadStationRemovalArgs(const registers& regs)
-            : pos(regs.ax, regs.cx, regs.di)
-            , rotation(regs.bh & 0x3)
-            , roadId(regs.dl & 0xF)
-            , index(regs.dh & 0x3)
-            , type(regs.bp & 0xF)
-        {
-        }
-
-        World::Pos3 pos;
-        uint8_t rotation;
-        uint8_t roadId;
-        uint8_t index;
-        uint8_t type;
-
-        explicit operator registers() const
-        {
-            registers regs;
-            regs.ax = pos.x;
-            regs.cx = pos.y;
-            regs.di = pos.z;
-            regs.bh = rotation;
-            regs.dl = roadId;
-            regs.dh = index;
-            regs.bp = type;
-            return regs;
-        }
-    };
-
     struct BuildingPlacementArgs
     {
         static constexpr auto command = GameCommand::createBuilding;
